Intégration de la logique de recherche des chemins à la classe OccupancyGrid
