+++ src/swarm_rescue/launcher.py
@@ -72,7 +72,7 @@ class Launcher:
         self.team_info = TeamInfo()
         self.eval_plan = EvalPlan()
 
-        eval_config = EvalConfig(map_type=MyMaptest2, nb_rounds=1)
+        eval_config = EvalConfig(map_type=MyMaptest3, nb_rounds=1)
         self.eval_plan.add(eval_config=eval_config)
 
         # eval_config = EvalConfig(map_type=MyMapIntermediate02)
